cleanup of some code of ex_test; merge websocket and web_utils
